// frontend/src/types/generated/models/DiagnosticReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiagnosticSummary } from "./DiagnosticSummary";
import type { FailedCaseSummary } from "./FailedCaseSummary";
import type { TurningPoint } from "./TurningPoint";

/**
 * 诊断报告
 */
export type DiagnosticReport = { taskId: string, taskName: string, status: string, summary: DiagnosticSummary, turningPoints: Array<TurningPoint>, improvementSuggestions: Array<string>, failedCases: Array<FailedCaseSummary>, };